Resolved comments and Add columns reusable component added
